Refactored representation class abstract/__init__ files
Representation abstract class location inconsistent with the rest of
things in SMQTK.
