+++ python/smqtk/representation/__init__.py
@@ -1,12 +1,6 @@
-from .code_index import get_index_types
-from .data_element_impl import get_data_element_impls
-from .data_set_impl import get_data_set_impls
-from .descriptor_element_impl import get_descriptor_element_impls
+from .code_index import CodeIndex, get_code_index_impls
+from .data_element import DataElement, get_data_element_impls
+from .data_set import DataSet, get_data_set_impls
+from .descriptor_element import DescriptorElement, get_descriptor_element_impls
 
-from .code_index import CodeIndex
-from .data_element_abstract import DataElement
-from .data_set_abstract import DataSet
-from .descriptor_element_abstract import DescriptorElement
 from .descriptor_element_factory import DescriptorElementFactory
-
-__author__ = 'purg'
