+++ python/smqtk/representation/data_element/file_element.py
