+++ python/smqtk/representation/data_element/memory_element.py
