+++ python/smqtk/representation/data_set/__init__.py
@@ -1,4 +1,3 @@
-
 import abc
 import collections
 import logging
@@ -111,3 +110,36 @@ class DataSet (collections.Set, Configurable):
 
         """
         return
+
+
+def get_data_set_impls(reload_modules=False):
+    """
+    Discover and return DataSet implementation classes found in the plugin
+    directory. Keys in the returned map are the names of the discovered classes
+    and the paired values are the actual class type objects.
+
+    We look for modules (directories or files) that start with and alphanumeric
+    character ('_' prefixed files/directories are hidden, but not recommended).
+
+    Within a module, we first look for a helper variable by the name
+    ``DATA_SET_CLASS``, which can either be a single class object or an
+    iterable of class objects, to be exported. If the variable is set to None,
+    we skip that module and do not import anything. If the variable is not
+    present, we look for a class by the same na e and casing as the module's
+    name. If neither are found, the module is skipped.
+
+    :param reload_modules: Explicitly reload discovered modules from source.
+    :type reload_modules: bool
+
+    :return: Map of discovered class objects of type ``DataSet`` whose keys
+        are the string names of the classes.
+    :rtype: dict[str, type]
+
+    """
+    import os
+    from smqtk.utils.plugin import get_plugins
+
+    this_dir = os.path.abspath(os.path.dirname(__file__))
+    helper_var = "DATA_SET_CLASS"
+    return get_plugins(__name__, this_dir, helper_var, DataSet, None,
+                       reload_modules)
