+++ python/smqtk/representation/descriptor_element/local_elements.py
