+++ python/smqtk/representation/descriptor_element/solr_element.py
