+++ python/smqtk/tests/algorithms/nn_index/test_SI_FLANN.py
@@ -5,7 +5,7 @@ import unittest
 import nose.tools as ntools
 import numpy
 
-from smqtk.representation.descriptor_element_impl.local_elements import \
+from smqtk.representation.descriptor_element.local_elements import \
     DescriptorMemoryElement
 from smqtk.algorithms.nn_index.flann import FlannNearestNeighborsIndex
 from smqtk.utils.file_utils import make_tempfile
