+++ python/smqtk/tests/algorithms/nn_index/test_SI_abstract.py
@@ -4,7 +4,7 @@ import mock
 import nose.tools as ntools
 import numpy
 
-from smqtk.representation.descriptor_element_impl.local_elements import \
+from smqtk.representation.descriptor_element.local_elements import \
     DescriptorMemoryElement
 from smqtk.algorithms.nn_index import NearestNeighborsIndex
