+++ python/smqtk/tests/representation/code_index/test_MemoryCodeIndex.py
@@ -6,7 +6,7 @@ import unittest
 import nose.tools as ntools
 import numpy
 
-from smqtk.representation.descriptor_element_impl.local_elements import DescriptorMemoryElement
+from smqtk.representation.descriptor_element.local_elements import DescriptorMemoryElement
 from smqtk.representation.code_index.memory import MemoryCodeIndex
 
 __author__ = 'purg'
