+++ python/smqtk/tests/representation/code_index/test_abstractCodeIndex.py
@@ -4,7 +4,7 @@ import mock
 import nose.tools as ntools
 import numpy
 
-from smqtk.representation.descriptor_element_impl.local_elements import DescriptorMemoryElement
+from smqtk.representation.descriptor_element.local_elements import DescriptorMemoryElement
 from smqtk.representation.code_index import CodeIndex
 
 __author__ = 'purg'
